refactor: split 5 large files into modular components for better maintainability
Split oversized files (578-1007 lines) into focused, single-responsibility modules:

• lib/translations/types.ts (1,007 lines) → 8 type modules
• lib/actions/users.ts (894 lines) → 5 specialized modules
• components/admin/ScheduledImportManager.tsx (680 lines) → 8 components
• lib/actions/csv-import/core.ts (578 lines) → 3 modules
• evaluation-summary/EvaluationSummaryClient.tsx (589 lines) → 7 modules

All splits maintain backward compatibility through index.ts re-exports.
Achieved 80-90% size reduction per file while following React/TypeScript best practices.
