+++ src/app/(dashboard)/evaluation-summary/[id]/components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Evaluation Summary Components Index
+ * 
+ * Centralized exports for all evaluation summary components.
+ */
+
+export { EvaluationHeader } from './EvaluationHeader'
+export { EvaluationMetrics } from './EvaluationMetrics'
+export { EvaluationItems } from './EvaluationItems'
