+++ src/lib/actions/users/crud.ts
@@ -1,52 +1,18 @@
+/**
+ * User Management CRUD Operations
+ * 
+ * Core create, read, update, and delete operations for user management
+ * including validation, security checks, and audit logging.
+ */
+
 'use server'
 
 import { revalidatePath } from 'next/cache'
 import { prisma } from '@/lib/prisma-client'
-import { auth } from '@/auth'
 import bcrypt from 'bcryptjs'
-import { z } from 'zod'
 import { auditUserManagement } from '@/lib/services/audit-service'
-
-
-// Validation schemas
-const userSchema = z.object({
-  name: z.string().min(1, 'Name is required'),
-  email: z.string().email('Invalid email').optional().or(z.literal('')),
-  username: z.string().min(1, 'Username is required').optional().or(z.literal('')),
-  role: z.enum(['hr', 'manager', 'employee'], { 
-    message: 'Role must be hr, manager, or employee'
-  }),
-  companyId: z.string().min(1, 'Company is required'),
-  managerId: z.string().optional().or(z.literal('')),
-  userType: z.enum(['office', 'operational']).default('office'),
-  department: z.string().optional().or(z.literal('')),
-  position: z.string().optional().or(z.literal('')),
-  employeeId: z.string().optional().or(z.literal('')),
-  password: z.string().min(6, 'Password must be at least 6 characters').optional(),
-  pinCode: z.string().optional().or(z.literal('')),
-  active: z.boolean().default(true)
-})
-
-const editUserSchema = userSchema.omit({ active: true }).extend({
-  password: z.string().min(6, 'Password must be at least 6 characters').optional().or(z.literal(''))
-})
-
-const changePasswordSchema = z.object({
-  currentPassword: z.string().min(1, 'Current password is required'),
-  newPassword: z.string().min(8, 'New password must be at least 8 characters'),
-  confirmPassword: z.string().min(1, 'Password confirmation is required')
-}).refine((data) => data.newPassword === data.confirmPassword, {
-  message: "Passwords don't match",
-  path: ["confirmPassword"]
-})
-
-async function requireHRAccess() {
-  const session = await auth()
-  if (!session?.user?.id || session.user.role !== 'hr') {
-    throw new Error('Access denied - HR role required')
-  }
-  return session.user
-}
+import { requireHRAccess } from './utils'
+import { userSchema, editUserSchema } from './schemas'
 
 export async function createUser(formData: FormData) {
   try {
@@ -426,92 +392,6 @@ export async function deleteUser(userId: string) {
   }
 }
 
-export async function changePassword(formData: FormData) {
-  try {
-    const session = await auth()
-    if (!session?.user?.id) {
-      return {
-        success: false,
-        message: 'Authentication required'
-      }
-    }
-
-    const rawData = {
-      currentPassword: formData.get('currentPassword'),
-      newPassword: formData.get('newPassword'),
-      confirmPassword: formData.get('confirmPassword')
-    }
-
-    const result = changePasswordSchema.safeParse(rawData)
-    
-    if (!result.success) {
-      return {
-        success: false,
-        errors: result.error.flatten().fieldErrors,
-        message: 'Validation failed'
-      }
-    }
-
-    const { currentPassword, newPassword } = result.data
-
-    // Get user's current password hash
-    const user = await prisma.user.findUnique({
-      where: { id: session.user.id },
-      select: { passwordHash: true }
-    })
-
-    if (!user?.passwordHash) {
-      return {
-        success: false,
-        message: 'No password set for this account'
-      }
-    }
-
-    // Verify current password
-    const isCurrentPasswordValid = await bcrypt.compare(currentPassword, user.passwordHash)
-    if (!isCurrentPasswordValid) {
-      return {
-        success: false,
-        message: 'Current password is incorrect',
-        errors: { currentPassword: ['Current password is incorrect'] }
-      }
-    }
-
-    // Hash new password
-    const newPasswordHash = await bcrypt.hash(newPassword, 12)
-
-    // Update password
-    await prisma.user.update({
-      where: { id: session.user.id },
-      data: { passwordHash: newPasswordHash }
-    })
-
-    // Audit log for password change
-    await auditUserManagement(
-      session.user.id,
-      session.user.role,
-      session.user.companyId,
-      'update',
-      session.user.id,
-      undefined,
-      { action: 'password_change', timestamp: new Date().toISOString() },
-      'Password changed by user'
-    )
-
-    revalidatePath('/settings')
-    return {
-      success: true,
-      message: 'Password updated successfully'
-    }
-  } catch (error) {
-    console.error('Error changing password:', error)
-    return {
-      success: false,
-      message: 'Failed to update password'
-    }
-  }
-}
-
 // Advanced function for force deleting users with evaluations (HR only)
 export async function forceDeleteUser(userId: string, reason: string) {
   try {
@@ -629,267 +509,6 @@ export async function forceDeleteUser(userId: string, reason: string) {
     return {
       success: false,
       message: 'Failed to delete user and evaluation data'
-    }
-  }
-}
-
-// Archive user with snapshot
-export async function archiveUser(userId: string, reason?: string) {
-  try {
-    const currentUser = await requireHRAccess()
-
-    // Get user to archive with complete context
-    const userToArchive = await prisma.user.findFirst({
-      where: {
-        id: userId,
-        companyId: currentUser.companyId
-      },
-      include: {
-        company: true,
-        manager: {
-          select: { name: true, email: true }
-        }
-      }
-    })
-
-    if (!userToArchive) {
-      return {
-        success: false,
-        messageKey: 'userNotFound',
-        message: 'User not found'
-      }
-    }
-
-    if (!userToArchive.active) {
-      return {
-        success: false,
-        message: 'User is already archived'
-      }
-    }
-
-    // Create snapshot data
-    const archiveSnapshot = {
-      archivedAt: new Date(),
-      archivedReason: reason || null,
-      archivedManagerName: userToArchive.manager?.name || null,
-      archivedManagerEmail: userToArchive.manager?.email || null,
-      archivedDepartment: userToArchive.department || null,
-      archivedPosition: userToArchive.position || null,
-      archivedCompanyName: userToArchive.company.name
-    }
-
-    // Archive the user
-    await prisma.user.update({
-      where: { id: userId },
-      data: {
-        active: false,
-        ...archiveSnapshot
-      }
-    })
-
-    // Audit the archive action
-    await auditUserManagement(
-      currentUser.id,
-      currentUser.role,
-      currentUser.companyId,
-      'archive',
-      userId,
-      { 
-        active: true, 
-        manager: userToArchive.manager?.name,
-        department: userToArchive.department 
-      },
-      {
-        active: false,
-        archivedAt: archiveSnapshot.archivedAt.toISOString(),
-        archivedReason: archiveSnapshot.archivedReason,
-        archivedManagerName: archiveSnapshot.archivedManagerName,
-        archivedManagerEmail: archiveSnapshot.archivedManagerEmail,
-        archivedDepartment: archiveSnapshot.archivedDepartment,
-        archivedPosition: archiveSnapshot.archivedPosition,
-        archivedCompanyName: archiveSnapshot.archivedCompanyName
-      },
-      reason ? `Archived: ${reason}` : 'User archived'
-    )
-
-    return {
-      success: true,
-      message: 'User archived successfully'
-    }
-  } catch (error) {
-    console.error('Error archiving user:', error)
-    return {
-      success: false,
-      message: 'Failed to archive user'
-    }
-  }
-}
-
-// Bulk archive users
-export async function bulkArchiveUsers(userIds: string[], reason?: string) {
-  try {
-    const currentUser = await requireHRAccess()
-
-    // Get all users to archive
-    const usersToArchive = await prisma.user.findMany({
-      where: {
-        id: { in: userIds },
-        companyId: currentUser.companyId,
-        active: true
-      },
-      include: {
-        company: true,
-        manager: {
-          select: { name: true, email: true }
-        }
-      }
-    })
-
-    if (usersToArchive.length === 0) {
-      return {
-        success: false,
-        message: 'No active users found to archive'
-      }
-    }
-
-    // Check for manager dependencies
-    const managersBeingArchived = usersToArchive.filter(user => user.role === 'manager')
-    const dependentCount = await prisma.user.count({
-      where: {
-        managerId: { in: managersBeingArchived.map(m => m.id) },
-        active: true
-      }
-    })
-
-    if (dependentCount > 0) {
-      return {
-        success: false,
-        message: `Cannot archive ${managersBeingArchived.length} manager(s) - they have ${dependentCount} active employee(s). Please reassign their reports first.`
-      }
-    }
-
-    const archiveTimestamp = new Date()
-
-    // Bulk update all users
-    const archivePromises = usersToArchive.map(user =>
-      prisma.user.update({
-        where: { id: user.id },
-        data: {
-          active: false,
-          archivedAt: archiveTimestamp,
-          archivedReason: reason || null,
-          archivedManagerName: user.manager?.name || null,
-          archivedManagerEmail: user.manager?.email || null,
-          archivedDepartment: user.department || null,
-          archivedPosition: user.position || null,
-          archivedCompanyName: user.company.name
-        }
-      })
-    )
-
-    await Promise.all(archivePromises)
-
-    // Audit the bulk archive action
-    await auditUserManagement(
-      currentUser.id,
-      currentUser.role,
-      currentUser.companyId,
-      'bulk_archive',
-      '', // No single entity ID for bulk action
-      {},
-      {
-        archivedUserIds: userIds,
-        archivedCount: usersToArchive.length,
-        reason: reason,
-        archivedAt: archiveTimestamp.toISOString()
-      },
-      reason ? `Bulk archived ${usersToArchive.length} users: ${reason}` : `Bulk archived ${usersToArchive.length} users`
-    )
-
-    return {
-      success: true,
-      message: `Successfully archived ${usersToArchive.length} user(s)`
-    }
-  } catch (error) {
-    console.error('Error bulk archiving users:', error)
-    return {
-      success: false,
-      message: 'Failed to archive users'
-    }
-  }
-}
-
-// Unarchive user (restore)
-export async function unarchiveUser(userId: string) {
-  try {
-    const currentUser = await requireHRAccess()
-
-    // Check if user exists and is archived
-    const userToUnarchive = await prisma.user.findFirst({
-      where: {
-        id: userId,
-        companyId: currentUser.companyId,
-        active: false
-      }
-    })
-
-    if (!userToUnarchive) {
-      return {
-        success: false,
-        message: 'Archived user not found'
-      }
-    }
-
-    // Store archive data for audit
-    const archiveData = {
-      archivedAt: userToUnarchive.archivedAt,
-      archivedReason: userToUnarchive.archivedReason,
-      archivedManagerName: userToUnarchive.archivedManagerName,
-      archivedDepartment: userToUnarchive.archivedDepartment
-    }
-
-    // Unarchive the user (clear archive data)
-    await prisma.user.update({
-      where: { id: userId },
-      data: {
-        active: true,
-        archivedAt: null,
-        archivedReason: null,
-        archivedManagerName: null,
-        archivedManagerEmail: null,
-        archivedDepartment: null,
-        archivedPosition: null,
-        archivedCompanyName: null
-      }
-    })
-
-    // Audit the unarchive action
-    await auditUserManagement(
-      currentUser.id,
-      currentUser.role,
-      currentUser.companyId,
-      'unarchive',
-      userId,
-      { 
-        active: false, 
-        archivedAt: archiveData.archivedAt ? archiveData.archivedAt.toISOString() : null,
-        archivedReason: archiveData.archivedReason,
-        archivedManagerName: archiveData.archivedManagerName,
-        archivedDepartment: archiveData.archivedDepartment
-      },
-      { active: true },
-      'User unarchived (restored)'
-    )
-
-    return {
-      success: true,
-      message: 'User unarchived successfully'
-    }
-  } catch (error) {
-    console.error('Error unarchiving user:', error)
-    return {
-      success: false,
-      message: 'Failed to unarchive user'
     }
   }
 }
