+++ src/lib/actions/users/index.ts
@@ -1,0 +1,21 @@
+/**
+ * User Management Actions Index
+ * 
+ * Centralized exports for all user management operations including
+ * CRUD operations, password management, archiving, and validation schemas.
+ */
+
+// Export all schemas and types
+export * from './schemas'
+
+// Export utility functions
+export * from './utils'
+
+// Export CRUD operations
+export * from './crud'
+
+// Export password management
+export * from './password'
+
+// Export archive management
+export * from './archive'
