--- src/lib/translations/types.ts
@@ -1,1008 +1,0 @@
-export type Language = 'en' | 'es'
-
-export interface StatusTranslations {
-  active: string
-  inactive: string
-  completed: string
-  pending: string
-  draft: string
-  draftReopened: string
-  submitted: string
-  approved: string
-  inProgress: string
-  notStarted: string
-  overdue: string
-  archived: string
-  closed: string
-}
-
-export interface RatingsTranslations {
-  outstanding: string
-  exceedsExpectations: string
-  meetsExpectations: string
-  belowExpectations: string
-  needsImprovement: string
-  notRated: string
-}
-
-export interface CommonTranslations {
-  loading: string
-  save: string
-  cancel: string
-  edit: string
-  delete: string
-  submit: string
-  next: string
-  previous: string
-  signOut: string
-  dashboard: string
-  back: string
-  yes: string
-  no: string
-  standardized: string
-  company: string
-  department: string
-  manager: string
-  createdBy: string
-  saving: string
-  approving: string
-  of: string
-  employees: string
-  // AI Features
-  improveWithAI: string
-  improving: string
-  aiImprovement: string
-  improveWithAITooltip: string
-  aiImprovementSuccess: string
-  aiImprovementFailed: string
-  dismissError: string
-  departments: string
-  unknownError: string
-  error: string
-  found: string
-  selected: string
-  unassigned: string
-  selectAll: string
-  deselectAll: string
-  noEmployeesFound: string
-  noDepartmentsFound: string
-  allDepartments: string
-  hide: string
-  show: string
-  view: string
-  approve: string
-  recall: string
-  unlock: string
-  you: string
-  remove: string
-  never: string
-  created: string
-  lastUsed: string
-  name: string
-  na: string
-  authenticating: string
-  authenticationSubtitle: string
-  welcomeBack: string
-  yourDashboard: string
-  takingYouTo: string
-  secondsRemaining: string
-  secondRemaining: string
-  versions: string
-  refine: string
-}
-
-export interface AuthTranslations {
-  signIn: string
-  signOut: string
-  emailOrUsername: string
-  password: string
-  signingIn: string
-  invalidCredentials: string
-  loginFailed: string
-  signInToAccount: string
-  demoCredentials: string
-  hr: string
-  manager: string
-  employee: string
-  worker: string
-  name: string
-  email: string
-  role: string
-}
-
-export interface NavigationTranslations {
-  performanceManagement: string
-  employeeEvaluations: string
-  myEvaluations: string
-  selectEmployee: string
-  employees: string
-  assignments: string
-  welcomeBack: string
-  performanceHistory: string
-  currentPeriod: string
-  settings: string
-  evaluationComplete: string
-  latestPerformanceEvaluation: string
-  reviewedBy: string
-  evaluatedOn: string
-  evaluatedBy: string
-  noEvaluationsYet: string
-  noPerformanceEvaluations: string
-  evaluationHistory: string
-  viewDetails: string
-  performanceSummary: string
-  awaitingYourApproval: string
-  evaluationInProgress: string
-  managerPreparingEvaluation: string
-  managerSubmittedEvaluation: string
-  averageRating: string
-  completedReviews: string
-  consistentlyExceeding: string
-  performanceOutstanding: string
-  performanceExceeding: string
-  performanceMeeting: string
-  performanceImproving: string
-  performanceNeedsWork: string
-}
-
-export interface EvaluationsTranslations {
-  evaluation: string
-  evaluateEmployee: string
-  employeeEvaluation: string
-  selectEmployeeToEvaluate: string
-  overallPerformance: string
-  overallRating: string
-  overallComments: string
-  submitEvaluation: string
-  submitForApproval: string
-  submitting: string
-  evaluationSubmittedSuccess: string
-  evaluationApprovedSuccess: string
-  awaitingEmployeeApproval: string
-  evaluationCompleted: string
-  unlocking: string
-  unlock: string
-  evaluationUnlocked: string
-  progress: string
-  of: string
-  itemsCompleted: string
-  okr: string
-  competency: string
-  ratePerformance: string
-  optionalComment: string
-  provideFeedback: string
-  addSpecificFeedback: string
-  teamSummary: string
-  pendingReviews: string
-  teamAverage: string
-  evaluationSubmitted: string
-  evaluationSubmittedDesc: string
-  redirecting: string
-  newCompanyItemsAdded: string
-  employeeNeedReEvaluation: string
-  employeesNeedReEvaluation: string
-  evaluationsAwaitingApproval: string
-  evaluationAwaitingApproval: string
-  comments: string
-  minimumCharacters: string
-  tapToRate: string
-  commentPlaceholder: string
-  commentGuidance: string
-}
-
-export interface SpeechTranslations {
-  startRecording: string
-  stopRecording: string
-  listening: string
-  tapToSpeak: string
-  microphoneAccess: string
-  networkError: string
-  noSpeechDetected: string
-  audioNotSupported: string
-  speakNow: string
-  speechNotSupported: string
-  transcriptionError: string
-  lengthExceeded: string
-}
-
-export interface BiometricTranslations {
-  setup: string
-  use: string
-  setupSuccess: string
-  loginSuccess: string
-  notSupported: string
-  cancelled: string
-  failed: string
-  httpsRequired: string
-  credentialExists: string
-  noCredentials: string
-  tapToSetup: string
-  tapToAuth: string
-  settingUp: string
-  authenticating: string
-  faceIdTouchId: string
-  touchId: string
-  fingerprint: string
-  biometric: string
-  userInfoRequired: string
-  authentication: string
-  settingsDescription: string
-  addNew: string
-  existingCredentials: string
-  confirmRemove: string
-  removeSuccess: string
-  faceId: string
-}
-
-export interface SettingsTranslations {
-  title: string
-  subtitle: string
-  profile: string
-  profileDescription: string
-  editProfile: string
-  changePassword: string
-  currentPassword: string
-  newPassword: string
-  confirmPassword: string
-  passwordsMatch: string
-  passwordUpdated: string
-  passwordSecurityNote: string
-  currentPasswordPlaceholder: string
-  newPasswordPlaceholder: string
-  confirmPasswordPlaceholder: string
-  noCredentials: string
-  noCredentialsDescription: string
-  failedToLoadCredentials: string
-  failedToRemoveCredential: string
-}
-
-export interface AssignmentsTranslations {
-  assignmentManager: string
-  manageOKRsCompetencies: string
-  companyWideItems: string
-  companyWideDescription: string
-  appliedToAllEmployees: string
-  selectEmployeesForBatch: string
-  selectEmployees: string
-  itemsAssigned: string
-  employeesSelected: string
-  assignToSelected: string
-  individualAssignments: string
-  individualDescription: string
-  comingSoon: string
-  createNewDepartmentItems: string
-  newOKR: string
-  newCompetency: string
-  create: string
-  evaluationDeadline: string
-  departmentLevelAssignments: string
-  departmentDescription: string
-  currentlyAssignedTo: string
-}
-
-export interface CompanyItemsTranslations {
-  title: string
-  subtitle: string
-  infoTitle: string
-  infoDescription: string
-  createNewTitle: string
-  newCompanyOKR: string
-  newCompanyCompetency: string
-  existingItemsTitle: string
-  itemsAppliedToAll: string
-  objective: string
-  competencyName: string
-  keyResults: string
-  description: string
-  activate: string
-  deactivate: string
-  active: string
-  inactive: string
-  createdBy: string
-  appliedToAllEmployees: string
-  noItemsTitle: string
-  noItemsDescription: string
-  companyOKR: string
-  companyCompetency: string
-  companyWide: string
-  confirmDeactivate: string
-  confirmActivate: string
-  deactivateWarning: string
-  activateWarning: string
-}
-
-export interface OKRsTranslations {
-  objective: string
-  keyResults: string
-}
-
-export interface DashboardTranslations {
-  hrDashboard: string
-  q1Reviews: string
-  newEvaluation: string
-  completionStatus: string
-  complete: string
-  completed: string
-  remaining: string
-  overdueEvaluations: string
-  pastDueDate: string
-  sendReminders: string
-  dueSoon: string
-  dueWithinDays: string
-  viewList: string
-  ratingDistribution: string
-  generalResults: string
-  outstanding: string
-  exceeds: string
-  meets: string
-  below: string
-  needsImprovement: string
-  quickActions: string
-  administrativeActions: string
-  generateReports: string
-  exportPDF: string
-  pdfExportCenter: string
-  exportCenter: string
-  selectReportsAndFormats: string
-  companyOverview: string
-  allEmployeesAllDepartments: string
-  departmentSummary: string
-  departmentBreakdown: string
-  topPerformers: string
-  highestRatedEmployees: string
-  needsAttention: string
-  employeesNeedingSupport: string
-  managerReports: string
-  reportsGroupedByManager: string
-  customSelection: string
-  selectSpecificEmployees: string
-  reportsSelected: string
-  exporting: string
-  exportSelected: string
-  exportError: string
-  evaluationNotFound: string
-  accessDenied: string
-  exportFailed: string
-  noEvaluationsFound: string
-  hrRoleRequired: string
-  managerOrHrRequired: string
-  allPeriods: string
-  manageUsers: string
-  archivedEmployees: string
-  startEvaluating: string
-  lastUpdated: string
-  newCycle: string
-  cycleName: string
-  startDate: string
-  endDate: string
-  createCycle: string
-  creating: string
-  overdueDrafts: string
-  pendingApprovals: string
-  daysPending: string
-  submittedBy: string
-  overdue3Days: string
-  viewAll: string
-  companyWideItems: string
-  auditLogs: string
-  auditDashboard: string
-  auditDescription: string
-  auditFilters: string
-  allActions: string
-  allTypes: string
-  clearFilters: string
-  showDetails: string
-  hideDetails: string
-  auditDetails: string
-  exportExcel: string
-  showingOf: string
-  noAuditLogs: string
-  timestamp: string
-  user: string
-  action: string
-  entity: string
-  details: string
-  ipAddress: string
-  reason: string
-  oldData: string
-  newData: string
-  metadata: string
-  cycles: string
-  performanceCycles: string
-  cycleManagement: string
-  activeCycle: string
-  closedCycle: string
-  archivedCycle: string
-  active: string
-  closed: string
-  archived: string
-  evaluationsText: string
-  itemsText: string
-  assessmentsText: string
-  manageCycles: string
-  cycleStatus: string
-  cycleClosedBy: string
-  cycleClosedAt: string
-  reopenCycle: string
-  closeCycle: string
-  archiveCycle: string
-  deleteCycle: string
-  confirmClose: string
-  confirmReopen: string
-  confirmArchive: string
-  confirmDelete: string
-  closeWarning: string
-  reopenWarning: string
-  deleteWarning: string
-  onlyOneActive: string
-  cannotDeleteData: string
-  dateRange: string
-  exportAllPDF: string
-  exportAllExcel: string
-  clearAll: string
-  departmentRatings: string
-  performanceInsights: string
-  totalEvaluations: string
-  completeDepartments: string
-  departmentsNeedAttention: string
-  swipeForDetails: string
-  noManager: string
-  notAssigned: string
-  employees: string
-  evaluated: string
-  pending: string
-  performanceDistribution: string
-  evaluations: string
-  outstandingShort: string
-  exceedsShort: string
-  meetsShort: string
-  belowShort: string
-  needsImprovementShort: string
-  needAttention: string
-  noCompletedEvaluations: string
-  evaluationsWillAppear: string
-  noDepartmentsFound: string
-  addEmployeesWithDepartments: string
-  employeesNeedReview: string
-  allEmployees: string
-  viewAllEmployees: string
-  pendingEvaluations: string
-  managePendingEvaluations: string
-  employeesWithoutEvaluations: string
-  startEvaluation: string
-  searchEmployees: string
-  searchEmployeesManagers: string
-  allEvaluationsCompleted: string
-  excellentWork: string
-  noManagerAssigned: string
-  viewResults: string
-  scheduleOneOnOnes: string
-  viewCompletedEvaluations: string
-  okrAverage: string
-  competencyAverage: string
-  totalAverage: string
-  objectives: string
-  skills: string
-  items: string
-  managerComments: string
-  feedback: string
-  companyLevel: string
-  departmentLevel: string
-  managerLevel: string
-  searchDepartments: string
-  allDepartments: string
-  allStatus: string
-  hasEvaluation: string
-  noEvaluation: string
-  partial: string
-  closedOn: string
-  performanceCycleClosedHR: string
-  performanceCycleClosedManager: string
-  performanceCycleArchived: string
-  performanceCycleRestricted: string
-  all: string
-  overdue: string
-  high: string
-  medium: string
-  low: string
-  highPriority: string
-  mediumPriority: string
-  lowPriority: string
-  department: string
-  manager: string
-  dueThisWeek: string
-  createdBy: string
-  deadlineSetBy: string
-  noDeadline: string
-  allEvaluationItems: string
-  itemsWithDeadlines: string
-  daysRemaining: string
-  dayRemaining: string
-  dayOverdue: string
-  dueToday: string
-  managersWithIssues: string
-  employeesBehind: string
-  overdueItems: string
-  managerEvaluationAccountability: string
-  managersWithEmployeesOverdue: string
-  allEmployeesWithOverdueEvaluations: string
-  employeesBehindOnEvaluations: string
-  noManagerIssuesFound: string
-  allEmployeesUpToDate: string
-  noOverdueEvaluations: string
-  employeesBehindEvaluations: string
-  overdueItemsCount: string
-  daysOverdue: string
-  switchToEmployeeList: string
-  switchToManagerGroups: string
-  
-  // Export Center Modal
-  reportSelected: string
-  reportsSelectedPlural: string
-  selectReportsToExport: string
-  selectEmployeesToExport: string
-  selectDepartmentsToExport: string
-  employeesSelected: string
-  departmentsSelected: string
-  selectReportsFirst: string
-  selectEmployeesFirst: string
-  selectDepartmentsFirst: string
-  exportFiles: string
-  exportFilesPlural: string
-  
-  // Administrative Actions Section
-  attentionRequired: string
-  deadlines: string
-  primaryManagement: string
-  systemAdministration: string
-  monitoringAnalysis: string
-  evaluationAssignments: string
-  companyAssignments: string
-  manageOkrsCompetencies: string
-  manageCompanyOkrsCompetencies: string
-  generateReportsAnalysis: string
-  totalEmployeesHeader: string
-  avgRatingHeader: string
-  noActivePerformanceCycle: string
-  
-  // Completion Status Card
-  clickToManagePending: string
-  completeStatus: string
-  inProgressStatus: string
-  
-  // Critical Actions
-  overdueEvaluationsCount: string
-  dueSoonCount: string
-  overdueText: string
-  daysOverdueText: string
-  viewAllOverdue: string
-  managerLabel: string
-  
-  // Department Ratings specific
-  comprehensiveAnalysisDescription: string
-  requireImmediateAttention: string
-  total: string
-  
-  // Pending Evaluations specific
-  employeeEvaluationManagement: string
-  evaluationsFinished: string
-  awaitingEvaluation: string
-  withPendingWork: string
-  employeeSingular: string
-  employeesPlural: string
-  awaitingEvaluationText: string
-  reportsTo: string
-  clearSearch: string
-  departmentsHeader: string
-  pendingText: string
-  
-  // Action badges for audit logs
-  actionCreate: string
-  actionUpdate: string
-  actionDelete: string
-  actionSubmit: string
-  actionApprove: string
-  actionUnlock: string
-  actionLogin: string
-  actionExport: string
-  actionImport: string
-}
-
-export interface UsersTranslations {
-  userManagement: string
-  totalUsers: string
-  importUsersCSV: string
-  uploadCSVFile: string
-  csvFormat: string
-  downloadExample: string
-  databaseManagement: string
-  advancedOperations: string
-  openPrismaStudio: string
-  searchUsers: string
-  allRoles: string
-  active: string
-  inactive: string
-  department: string
-  manager: string
-  manages: string
-  evaluationsReceived: string
-  noUsersFound: string
-  uploadingProcessing: string
-  successfullyImported: string
-  uploadFailed: string
-  basicExampleCSV: string
-  advancedExampleCSV: string
-  usesPersonIdMatching: string
-  showsBothIdOptions: string
-  fieldExplanations: string
-  employeeIdExplanation: string
-  personIdExplanation: string
-  managerPersonIdExplanation: string
-  managerEmployeeIdExplanation: string
-  advanced: string
-  advancedAdmin: string
-  dangerZone: string
-  resetDatabase: string
-  resetDatabaseTitle: string
-  resetDatabaseDescription: string
-  resetDatabaseWarning: string
-  resetDatabaseConfirm: string
-  resetDatabaseButton: string
-  resetSuccess: string
-  resetFailed: string
-  addUser: string
-  editUser: string
-  deleteUser: string
-  exportUsers: string
-  exportUsersDescription: string
-  exportExcel: string
-  exporting: string
-  exportSuccess: string
-  exportFailed: string
-  name: string
-  email: string
-  username: string
-  role: string
-  company: string
-  employeeId: string
-  position: string
-  personID: string
-  userType: string
-  loginMethod: string
-  shift: string
-  activeUser: string
-  password: string
-  newPassword: string
-  pinCode: string
-  requiresPinOnly: string
-  officeWorker: string
-  operationalWorker: string
-  selectRole: string
-  selectCompany: string
-  noManager: string
-  createUser: string
-  updateUser: string
-  leaveEmptyToKeep: string
-  fourDigitPin: string
-  deleteConfirmMessage: string
-  actionCannotBeUndone: string
-  deleting: string
-  deleteUserWarning: string
-  manageUsersDescription: string
-  clearSearch: string
-  usersCount: string
-  reportsTo: string
-  reports: string
-  evaluations: string
-  editUserTitle: string
-  deleteUserTitle: string
-  enterPassword: string
-  namePlaceholder: string
-  emailPlaceholder: string
-  usernamePlaceholder: string
-  departmentPlaceholder: string
-  employeeIdPlaceholder: string
-  positionPlaceholder: string
-  personIdPlaceholder: string
-  shiftPlaceholder: string
-  csvImportWorkflow: string
-  step: string
-  chooseFileOrDrag: string
-  selectCSVFileDescription: string
-  importGuide: string
-  downloadTemplate: string
-  previewImport: string
-  analyzing: string
-  configureImport: string
-  executeImport: string
-  importing: string
-  viewHistory: string
-  importHistory: string
-  backToImport: string
-  clickToChangeFile: string
-  csvFilesUpTo: string
-  manualImport: string
-  scheduledImports: string
-  csvRequirements: string
-  requiredFields: string
-  officeUsers: string
-  operationalUsers: string
-  managers: string
-  encoding: string
-  automateCSVImports: string
-  createSchedule: string
-  editScheduledImport: string
-  createNewScheduledImport: string
-  configureAutomatedImports: string
-  importName: string
-  description: string
-  schedule: string
-  frequency: string
-  time: string
-  timezone: string
-  dataSource: string
-  sourceType: string
-  authenticationOptional: string
-  lastRun: string
-  nextRun: string
-  noScheduledImports: string
-  createFirstScheduleDescription: string
-  createFirstSchedule: string
-  executeNow: string
-  pause: string
-  enable: string
-  update: string
-  create: string
-  failedToLoadScheduledImports: string
-  fillAllRequiredFields: string
-  scheduledImportCreated: string
-  failedToCreateScheduledImport: string
-  scheduledImportUpdated: string
-  failedToUpdateScheduledImport: string
-  scheduledImportDeleted: string
-  failedToDeleteScheduledImport: string
-  importExecutedSuccessfully: string
-  failedToExecuteScheduledImport: string
-  
-  // Additional keys for refactored components
-  goBack: string
-  signOut: string
-  searchUsersByNameEmail: string
-  user: string
-  users: string
-  found: string
-  total: string
-  managersLabel: string
-  employees: string
-  tryAdjustingSearch: string
-  noUsersCreated: string
-  noContact: string
-  noCompany: string
-  id: string
-  report: string
-  evaluation: string
-  delete: string
-  enterFullName: string
-  enterEmailAddress: string
-  enterUsername: string
-  enterDepartment: string
-  enterPosition: string
-  enterEmployeeId: string
-  enterPersonId: string
-  qrCode: string
-  enterShift: string
-  enterFourDigitPin: string
-  cancel: string
-  saving: string
-  closeModal: string
-  updateUserInformation: string
-  createNewUserAccount: string
-  areYouSureDelete: string
-  thisWillPermanently: string
-  deleteAllUserData: string
-  removeEvaluationRecords: string
-  unassignDirectReports: string
-  
-  // CSV Error Messages
-  usersWithErrors: string
-  totalErrors: string
-  errors: string
-  error: string
-  rowNumber: string
-  noIdentifier: string
-  dataContext: string
-  noDept: string
-  moreUsersWithErrors: string
-  validationErrorsWillBeSkipped: string
-  commonIssuesToCheck: string
-  missingRequiredFields: string
-  invalidPasswordFormat: string
-  managerReferencesNotExist: string
-  invalidEmailFormats: string
-  bomCharacterIssue: string
-  
-  // CSV Preview Step
-  previewResults: string
-  reviewAnalysisBeforeImporting: string
-  totalLabel: string
-  validLabel: string
-  newLabel: string
-  updatesLabel: string
-  validUsersPreview: string
-  newBadge: string
-  updBadge: string
-  andMoreValidUsers: string
-  back: string
-  
-  // CSV Configuration Step
-  configureImportOptions: string
-  chooseFieldsAndBehavior: string
-  importActions: string
-  createNewUsers: string
-  importUsersDontExist: string
-  updateExistingUsers: string
-  updateExistingUsersDesc: string
-  fieldsToImportUpdate: string
-  requiredFieldsAlwaysImported: string
-  errorHandling: string
-  skipRowsWithErrors: string
-  continueImportEvenFail: string
-  continueOnValidationErrors: string
-  processValidRowsEvenIssues: string
-  autoFixWeakPasswords: string
-  autoStrengthenPasswords: string
-  performanceOptions: string
-  enableBatchProcessing: string
-  processLargeImportsBatches: string
-  batchSizeUsers: string
-  importSummary: string
-  actionsLabel: string
-  createNew: string
-  updateExisting: string
-  fieldsSelected: string
-  errorHandlingLabel: string
-  skipErrors: string
-  stopOnErrors: string
-  performanceLabel: string
-  batchProcessing: string
-  standardProcessing: string
-  backToPreview: string
-  
-  // CSV Execution Step
-  importCompleted: string
-  importFailed: string
-  usersCreated: string
-  usersUpdated: string
-  failed: string
-  importCompletedSuccessfully: string
-  usersWereCreated: string
-  usersFailedToImport: string
-  importPartiallyCompleted: string
-  someUsersProcessed: string
-  importFailedMessage: string
-  importErrors: string
-  downloadErrorReport: string
-  andMoreErrors: string
-  recoverableErrors: string
-  suggestedFix: string
-  canRetry: string
-  andMoreRecoverableErrors: string
-  criticalErrors: string
-  requiresAdminAction: string
-  importAnotherFile: string
-  viewUsers: string
-  nextSteps: string
-  reviewImportedUsers: string
-  verifyManagerRelationships: string
-  testLoginCredentials: string
-  updateAdditionalSettings: string
-  partialSuccessActionRequired: string
-  reviewAndFixErrors: string
-  reimportCorrectedFile: string
-  checkImportedUsers: string
-  considerErrorRecovery: string
-  importErrorReport: string
-  generated: string
-  summaryLabel: string
-  totalErrorsReport: string
-  createdUsersReport: string
-  updatedUsersReport: string
-  failedUsersReport: string
-  errorsLabel: string
-  
-  // Retry Functionality
-  retryThis: string
-  retryAll: string
-  retryingErrors: string
-  autoFixAvailable: string
-  manualFixRequired: string
-  applyAutoFixes: string
-  manuallyFix: string
-  retryComplete: string
-  retryFailed: string
-  retriedRows: string
-  stillFailed: string
-  allErrorsFixed: string
-  someErrorsRemain: string
-  retryInProgress: string
-  processingFixes: string
-
-  // User deletion error messages
-  userNotFound: string
-  cannotDeleteUserManagesEmployees: string
-  cannotDeleteUserHasEvaluations: string
-  cannotDeleteDueToConstraints: string
-  userDeletedSuccessfully: string
-  failedToDeleteUser: string
-  userCreatedSuccessfully: string
-  failedToCreateUser: string
-  userUpdatedSuccessfully: string
-  failedToUpdateUser: string
-  userWithEmailExists: string
-  validationFailed: string
-
-  // Evaluation details for deletion errors
-  evaluationsAsEmployee: string
-  evaluationsAsManager: string
-  evaluationAssignments: string
-  partialAssessments: string
-
-  // Archive functionality
-  archivedEmployees: string
-  viewArchivedEmployees: string
-  archiveEmployee: string
-  archiveSelectedEmployees: string
-  unarchiveEmployee: string
-  archiveReason: string
-  archiveEmployeeTitle: string
-  archiveEmployeeDescription: string
-  archiveReasonPlaceholder: string
-  archiveReasonNote: string
-  archiving: string
-  archivingEmployee: string
-  archivingSelectedEmployees: string
-  close: string
-  unarchivingEmployee: string
-  employeeArchivedSuccessfully: string
-  employeesArchivedSuccessfully: string
-  employeeUnarchivedSuccessfully: string
-  failedToArchiveEmployee: string
-  failedToArchiveEmployees: string
-  failedToUnarchiveEmployee: string
-  cannotArchiveManagerWithActiveReports: string
-  cannotArchiveManager: string
-  cannotArchiveSelf: string
-  activeEmployees: string
-  noDataCanBeDeleted: string
-  archivedOn: string
-  archivedBy: string
-  archiveDetails: string
-  archivedManager: string
-  archivedDepartment: string
-  archivedPosition: string
-  noArchivedEmployees: string
-  noArchivedEmployeesDescription: string
-  noArchivedEmployeesFound: string
-  noArchivedEmployeesFoundDescription: string
-  evaluationHistory: string
-  showEvaluationHistory: string
-  hideEvaluationHistory: string
-  noEvaluations: string
-}
-
-export interface Translations {
-  status: StatusTranslations
-  ratings: RatingsTranslations
-  common: CommonTranslations
-  auth: AuthTranslations
-  nav: NavigationTranslations
-  evaluations: EvaluationsTranslations
-  speech: SpeechTranslations
-  biometric: BiometricTranslations
-  settings: SettingsTranslations
-  assignments: AssignmentsTranslations
-  companyItems: CompanyItemsTranslations
-  okrs: OKRsTranslations
-  dashboard: DashboardTranslations
-  users: UsersTranslations
-}
