+++ src/lib/translations/types/auth.ts
@@ -1,0 +1,60 @@
+/**
+ * Authentication Translation Types
+ * 
+ * Defines translation types for authentication-related UI elements,
+ * including login forms, user roles, and authentication states.
+ */
+
+export interface AuthTranslations {
+  signIn: string
+  signOut: string
+  emailOrUsername: string
+  password: string
+  signingIn: string
+  invalidCredentials: string
+  loginFailed: string
+  signInToAccount: string
+  demoCredentials: string
+  hr: string
+  manager: string
+  employee: string
+  worker: string
+  name: string
+  email: string
+  role: string
+}
+
+export interface NavigationTranslations {
+  performanceManagement: string
+  employeeEvaluations: string
+  myEvaluations: string
+  selectEmployee: string
+  employees: string
+  assignments: string
+  welcomeBack: string
+  performanceHistory: string
+  currentPeriod: string
+  settings: string
+  evaluationComplete: string
+  latestPerformanceEvaluation: string
+  reviewedBy: string
+  evaluatedOn: string
+  evaluatedBy: string
+  noEvaluationsYet: string
+  noPerformanceEvaluations: string
+  evaluationHistory: string
+  viewDetails: string
+  performanceSummary: string
+  awaitingYourApproval: string
+  evaluationInProgress: string
+  managerPreparingEvaluation: string
+  managerSubmittedEvaluation: string
+  averageRating: string
+  completedReviews: string
+  consistentlyExceeding: string
+  performanceOutstanding: string
+  performanceExceeding: string
+  performanceMeeting: string
+  performanceImproving: string
+  performanceNeedsWork: string
+}
