+++ src/lib/translations/types/common.ts
@@ -1,0 +1,73 @@
+/**
+ * Common Translation Types
+ * 
+ * Defines translation types for commonly used UI elements, actions, and messages
+ * that appear throughout the application. Includes AI-related translations and
+ * general interaction labels.
+ */
+
+export interface CommonTranslations {
+  loading: string
+  save: string
+  cancel: string
+  edit: string
+  delete: string
+  submit: string
+  next: string
+  previous: string
+  signOut: string
+  dashboard: string
+  back: string
+  yes: string
+  no: string
+  standardized: string
+  company: string
+  department: string
+  manager: string
+  createdBy: string
+  saving: string
+  approving: string
+  of: string
+  employees: string
+  // AI Features
+  improveWithAI: string
+  improving: string
+  aiImprovement: string
+  improveWithAITooltip: string
+  aiImprovementSuccess: string
+  aiImprovementFailed: string
+  dismissError: string
+  departments: string
+  unknownError: string
+  error: string
+  found: string
+  selected: string
+  unassigned: string
+  selectAll: string
+  deselectAll: string
+  noEmployeesFound: string
+  noDepartmentsFound: string
+  allDepartments: string
+  hide: string
+  show: string
+  view: string
+  approve: string
+  recall: string
+  unlock: string
+  you: string
+  remove: string
+  never: string
+  created: string
+  lastUsed: string
+  name: string
+  na: string
+  authenticating: string
+  authenticationSubtitle: string
+  welcomeBack: string
+  yourDashboard: string
+  takingYouTo: string
+  secondsRemaining: string
+  secondRemaining: string
+  versions: string
+  refine: string
+}
