+++ src/lib/translations/types/dashboard.ts
@@ -1,0 +1,299 @@
+/**
+ * Dashboard Translation Types
+ * 
+ * Defines translation types for the HR dashboard interface, including
+ * performance analytics, evaluation management, administrative actions,
+ * audit logs, and reporting features.
+ */
+
+export interface DashboardTranslations {
+  hrDashboard: string
+  q1Reviews: string
+  newEvaluation: string
+  completionStatus: string
+  complete: string
+  completed: string
+  remaining: string
+  overdueEvaluations: string
+  pastDueDate: string
+  sendReminders: string
+  dueSoon: string
+  dueWithinDays: string
+  viewList: string
+  ratingDistribution: string
+  generalResults: string
+  outstanding: string
+  exceeds: string
+  meets: string
+  below: string
+  needsImprovement: string
+  quickActions: string
+  administrativeActions: string
+  generateReports: string
+  exportPDF: string
+  pdfExportCenter: string
+  exportCenter: string
+  selectReportsAndFormats: string
+  companyOverview: string
+  allEmployeesAllDepartments: string
+  departmentSummary: string
+  departmentBreakdown: string
+  topPerformers: string
+  highestRatedEmployees: string
+  needsAttention: string
+  employeesNeedingSupport: string
+  managerReports: string
+  reportsGroupedByManager: string
+  customSelection: string
+  selectSpecificEmployees: string
+  reportsSelected: string
+  exporting: string
+  exportSelected: string
+  exportError: string
+  evaluationNotFound: string
+  accessDenied: string
+  exportFailed: string
+  noEvaluationsFound: string
+  hrRoleRequired: string
+  managerOrHrRequired: string
+  allPeriods: string
+  manageUsers: string
+  archivedEmployees: string
+  startEvaluating: string
+  lastUpdated: string
+  newCycle: string
+  cycleName: string
+  startDate: string
+  endDate: string
+  createCycle: string
+  creating: string
+  overdueDrafts: string
+  pendingApprovals: string
+  daysPending: string
+  submittedBy: string
+  overdue3Days: string
+  viewAll: string
+  companyWideItems: string
+  auditLogs: string
+  auditDashboard: string
+  auditDescription: string
+  auditFilters: string
+  allActions: string
+  allTypes: string
+  clearFilters: string
+  showDetails: string
+  hideDetails: string
+  auditDetails: string
+  exportExcel: string
+  showingOf: string
+  noAuditLogs: string
+  timestamp: string
+  user: string
+  action: string
+  entity: string
+  details: string
+  ipAddress: string
+  reason: string
+  oldData: string
+  newData: string
+  metadata: string
+  cycles: string
+  performanceCycles: string
+  cycleManagement: string
+  activeCycle: string
+  closedCycle: string
+  archivedCycle: string
+  active: string
+  closed: string
+  archived: string
+  evaluationsText: string
+  itemsText: string
+  assessmentsText: string
+  manageCycles: string
+  cycleStatus: string
+  cycleClosedBy: string
+  cycleClosedAt: string
+  reopenCycle: string
+  closeCycle: string
+  archiveCycle: string
+  deleteCycle: string
+  confirmClose: string
+  confirmReopen: string
+  confirmArchive: string
+  confirmDelete: string
+  closeWarning: string
+  reopenWarning: string
+  deleteWarning: string
+  onlyOneActive: string
+  cannotDeleteData: string
+  dateRange: string
+  exportAllPDF: string
+  exportAllExcel: string
+  clearAll: string
+  departmentRatings: string
+  performanceInsights: string
+  totalEvaluations: string
+  completeDepartments: string
+  departmentsNeedAttention: string
+  swipeForDetails: string
+  noManager: string
+  notAssigned: string
+  employees: string
+  evaluated: string
+  pending: string
+  performanceDistribution: string
+  evaluations: string
+  outstandingShort: string
+  exceedsShort: string
+  meetsShort: string
+  belowShort: string
+  needsImprovementShort: string
+  needAttention: string
+  noCompletedEvaluations: string
+  evaluationsWillAppear: string
+  noDepartmentsFound: string
+  addEmployeesWithDepartments: string
+  employeesNeedReview: string
+  allEmployees: string
+  viewAllEmployees: string
+  pendingEvaluations: string
+  managePendingEvaluations: string
+  employeesWithoutEvaluations: string
+  startEvaluation: string
+  searchEmployees: string
+  searchEmployeesManagers: string
+  allEvaluationsCompleted: string
+  excellentWork: string
+  noManagerAssigned: string
+  viewResults: string
+  scheduleOneOnOnes: string
+  viewCompletedEvaluations: string
+  okrAverage: string
+  competencyAverage: string
+  totalAverage: string
+  objectives: string
+  skills: string
+  items: string
+  managerComments: string
+  feedback: string
+  companyLevel: string
+  departmentLevel: string
+  managerLevel: string
+  searchDepartments: string
+  allDepartments: string
+  allStatus: string
+  hasEvaluation: string
+  noEvaluation: string
+  partial: string
+  closedOn: string
+  performanceCycleClosedHR: string
+  performanceCycleClosedManager: string
+  performanceCycleArchived: string
+  performanceCycleRestricted: string
+  all: string
+  overdue: string
+  high: string
+  medium: string
+  low: string
+  highPriority: string
+  mediumPriority: string
+  lowPriority: string
+  department: string
+  manager: string
+  dueThisWeek: string
+  createdBy: string
+  deadlineSetBy: string
+  noDeadline: string
+  allEvaluationItems: string
+  itemsWithDeadlines: string
+  daysRemaining: string
+  dayRemaining: string
+  dayOverdue: string
+  dueToday: string
+  managersWithIssues: string
+  employeesBehind: string
+  overdueItems: string
+  managerEvaluationAccountability: string
+  managersWithEmployeesOverdue: string
+  allEmployeesWithOverdueEvaluations: string
+  employeesBehindOnEvaluations: string
+  noManagerIssuesFound: string
+  allEmployeesUpToDate: string
+  noOverdueEvaluations: string
+  employeesBehindEvaluations: string
+  overdueItemsCount: string
+  daysOverdue: string
+  switchToEmployeeList: string
+  switchToManagerGroups: string
+  
+  // Export Center Modal
+  reportSelected: string
+  reportsSelectedPlural: string
+  selectReportsToExport: string
+  selectEmployeesToExport: string
+  selectDepartmentsToExport: string
+  employeesSelected: string
+  departmentsSelected: string
+  selectReportsFirst: string
+  selectEmployeesFirst: string
+  selectDepartmentsFirst: string
+  exportFiles: string
+  exportFilesPlural: string
+  
+  // Administrative Actions Section
+  attentionRequired: string
+  deadlines: string
+  primaryManagement: string
+  systemAdministration: string
+  monitoringAnalysis: string
+  evaluationAssignments: string
+  companyAssignments: string
+  manageOkrsCompetencies: string
+  manageCompanyOkrsCompetencies: string
+  generateReportsAnalysis: string
+  totalEmployeesHeader: string
+  avgRatingHeader: string
+  noActivePerformanceCycle: string
+  
+  // Completion Status Card
+  clickToManagePending: string
+  completeStatus: string
+  inProgressStatus: string
+  
+  // Critical Actions
+  overdueEvaluationsCount: string
+  dueSoonCount: string
+  overdueText: string
+  daysOverdueText: string
+  viewAllOverdue: string
+  managerLabel: string
+  
+  // Department Ratings specific
+  comprehensiveAnalysisDescription: string
+  requireImmediateAttention: string
+  total: string
+  
+  // Pending Evaluations specific
+  employeeEvaluationManagement: string
+  evaluationsFinished: string
+  awaitingEvaluation: string
+  withPendingWork: string
+  employeeSingular: string
+  employeesPlural: string
+  awaitingEvaluationText: string
+  reportsTo: string
+  clearSearch: string
+  departmentsHeader: string
+  pendingText: string
+  
+  // Action badges for audit logs
+  actionCreate: string
+  actionUpdate: string
+  actionDelete: string
+  actionSubmit: string
+  actionApprove: string
+  actionUnlock: string
+  actionLogin: string
+  actionExport: string
+  actionImport: string
+}
