+++ src/lib/translations/types/evaluations.ts
@@ -1,0 +1,52 @@
+/**
+ * Evaluations Translation Types
+ * 
+ * Defines translation types for the performance evaluation system,
+ * including evaluation forms, rating scales, progress tracking, and
+ * evaluation workflow states.
+ */
+
+export interface EvaluationsTranslations {
+  evaluation: string
+  evaluateEmployee: string
+  employeeEvaluation: string
+  selectEmployeeToEvaluate: string
+  overallPerformance: string
+  overallRating: string
+  overallComments: string
+  submitEvaluation: string
+  submitForApproval: string
+  submitting: string
+  evaluationSubmittedSuccess: string
+  evaluationApprovedSuccess: string
+  awaitingEmployeeApproval: string
+  evaluationCompleted: string
+  unlocking: string
+  unlock: string
+  evaluationUnlocked: string
+  progress: string
+  of: string
+  itemsCompleted: string
+  okr: string
+  competency: string
+  ratePerformance: string
+  optionalComment: string
+  provideFeedback: string
+  addSpecificFeedback: string
+  teamSummary: string
+  pendingReviews: string
+  teamAverage: string
+  evaluationSubmitted: string
+  evaluationSubmittedDesc: string
+  redirecting: string
+  newCompanyItemsAdded: string
+  employeeNeedReEvaluation: string
+  employeesNeedReEvaluation: string
+  evaluationsAwaitingApproval: string
+  evaluationAwaitingApproval: string
+  comments: string
+  minimumCharacters: string
+  tapToRate: string
+  commentPlaceholder: string
+  commentGuidance: string
+}
