+++ src/lib/translations/types/users.ts
@@ -1,0 +1,389 @@
+/**
+ * Users Translation Types
+ * 
+ * Defines translation types for user management functionality including
+ * user administration, CSV import/export, employee archiving, scheduled
+ * imports, and comprehensive user lifecycle management.
+ */
+
+export interface UsersTranslations {
+  userManagement: string
+  totalUsers: string
+  importUsersCSV: string
+  uploadCSVFile: string
+  csvFormat: string
+  downloadExample: string
+  databaseManagement: string
+  advancedOperations: string
+  openPrismaStudio: string
+  searchUsers: string
+  allRoles: string
+  active: string
+  inactive: string
+  department: string
+  manager: string
+  manages: string
+  evaluationsReceived: string
+  noUsersFound: string
+  uploadingProcessing: string
+  successfullyImported: string
+  uploadFailed: string
+  basicExampleCSV: string
+  advancedExampleCSV: string
+  usesPersonIdMatching: string
+  showsBothIdOptions: string
+  fieldExplanations: string
+  employeeIdExplanation: string
+  personIdExplanation: string
+  managerPersonIdExplanation: string
+  managerEmployeeIdExplanation: string
+  advanced: string
+  advancedAdmin: string
+  dangerZone: string
+  resetDatabase: string
+  resetDatabaseTitle: string
+  resetDatabaseDescription: string
+  resetDatabaseWarning: string
+  resetDatabaseConfirm: string
+  resetDatabaseButton: string
+  resetSuccess: string
+  resetFailed: string
+  addUser: string
+  editUser: string
+  deleteUser: string
+  exportUsers: string
+  exportUsersDescription: string
+  exportExcel: string
+  exporting: string
+  exportSuccess: string
+  exportFailed: string
+  name: string
+  email: string
+  username: string
+  role: string
+  company: string
+  employeeId: string
+  position: string
+  personID: string
+  userType: string
+  loginMethod: string
+  shift: string
+  activeUser: string
+  password: string
+  newPassword: string
+  pinCode: string
+  requiresPinOnly: string
+  officeWorker: string
+  operationalWorker: string
+  selectRole: string
+  selectCompany: string
+  noManager: string
+  createUser: string
+  updateUser: string
+  leaveEmptyToKeep: string
+  fourDigitPin: string
+  deleteConfirmMessage: string
+  actionCannotBeUndone: string
+  deleting: string
+  deleteUserWarning: string
+  manageUsersDescription: string
+  clearSearch: string
+  usersCount: string
+  reportsTo: string
+  reports: string
+  evaluations: string
+  editUserTitle: string
+  deleteUserTitle: string
+  enterPassword: string
+  namePlaceholder: string
+  emailPlaceholder: string
+  usernamePlaceholder: string
+  departmentPlaceholder: string
+  employeeIdPlaceholder: string
+  positionPlaceholder: string
+  personIdPlaceholder: string
+  shiftPlaceholder: string
+  csvImportWorkflow: string
+  step: string
+  chooseFileOrDrag: string
+  selectCSVFileDescription: string
+  importGuide: string
+  downloadTemplate: string
+  previewImport: string
+  analyzing: string
+  configureImport: string
+  executeImport: string
+  importing: string
+  viewHistory: string
+  importHistory: string
+  backToImport: string
+  clickToChangeFile: string
+  csvFilesUpTo: string
+  manualImport: string
+  scheduledImports: string
+  csvRequirements: string
+  requiredFields: string
+  officeUsers: string
+  operationalUsers: string
+  managers: string
+  encoding: string
+  automateCSVImports: string
+  createSchedule: string
+  editScheduledImport: string
+  createNewScheduledImport: string
+  configureAutomatedImports: string
+  importName: string
+  description: string
+  schedule: string
+  frequency: string
+  time: string
+  timezone: string
+  dataSource: string
+  sourceType: string
+  authenticationOptional: string
+  lastRun: string
+  nextRun: string
+  noScheduledImports: string
+  createFirstScheduleDescription: string
+  createFirstSchedule: string
+  executeNow: string
+  pause: string
+  enable: string
+  update: string
+  create: string
+  failedToLoadScheduledImports: string
+  fillAllRequiredFields: string
+  scheduledImportCreated: string
+  failedToCreateScheduledImport: string
+  scheduledImportUpdated: string
+  failedToUpdateScheduledImport: string
+  scheduledImportDeleted: string
+  failedToDeleteScheduledImport: string
+  importExecutedSuccessfully: string
+  failedToExecuteScheduledImport: string
+  
+  // Additional keys for refactored components
+  goBack: string
+  signOut: string
+  searchUsersByNameEmail: string
+  user: string
+  users: string
+  found: string
+  total: string
+  managersLabel: string
+  employees: string
+  tryAdjustingSearch: string
+  noUsersCreated: string
+  noContact: string
+  noCompany: string
+  id: string
+  report: string
+  evaluation: string
+  delete: string
+  enterFullName: string
+  enterEmailAddress: string
+  enterUsername: string
+  enterDepartment: string
+  enterPosition: string
+  enterEmployeeId: string
+  enterPersonId: string
+  qrCode: string
+  enterShift: string
+  enterFourDigitPin: string
+  cancel: string
+  saving: string
+  closeModal: string
+  updateUserInformation: string
+  createNewUserAccount: string
+  areYouSureDelete: string
+  thisWillPermanently: string
+  deleteAllUserData: string
+  removeEvaluationRecords: string
+  unassignDirectReports: string
+  
+  // CSV Error Messages
+  usersWithErrors: string
+  totalErrors: string
+  errors: string
+  error: string
+  rowNumber: string
+  noIdentifier: string
+  dataContext: string
+  noDept: string
+  moreUsersWithErrors: string
+  validationErrorsWillBeSkipped: string
+  commonIssuesToCheck: string
+  missingRequiredFields: string
+  invalidPasswordFormat: string
+  managerReferencesNotExist: string
+  invalidEmailFormats: string
+  bomCharacterIssue: string
+  
+  // CSV Preview Step
+  previewResults: string
+  reviewAnalysisBeforeImporting: string
+  totalLabel: string
+  validLabel: string
+  newLabel: string
+  updatesLabel: string
+  validUsersPreview: string
+  newBadge: string
+  updBadge: string
+  andMoreValidUsers: string
+  back: string
+  
+  // CSV Configuration Step
+  configureImportOptions: string
+  chooseFieldsAndBehavior: string
+  importActions: string
+  createNewUsers: string
+  importUsersDontExist: string
+  updateExistingUsers: string
+  updateExistingUsersDesc: string
+  fieldsToImportUpdate: string
+  requiredFieldsAlwaysImported: string
+  errorHandling: string
+  skipRowsWithErrors: string
+  continueImportEvenFail: string
+  continueOnValidationErrors: string
+  processValidRowsEvenIssues: string
+  autoFixWeakPasswords: string
+  autoStrengthenPasswords: string
+  performanceOptions: string
+  enableBatchProcessing: string
+  processLargeImportsBatches: string
+  batchSizeUsers: string
+  importSummary: string
+  actionsLabel: string
+  createNew: string
+  updateExisting: string
+  fieldsSelected: string
+  errorHandlingLabel: string
+  skipErrors: string
+  stopOnErrors: string
+  performanceLabel: string
+  batchProcessing: string
+  standardProcessing: string
+  backToPreview: string
+  
+  // CSV Execution Step
+  importCompleted: string
+  importFailed: string
+  usersCreated: string
+  usersUpdated: string
+  failed: string
+  importCompletedSuccessfully: string
+  usersWereCreated: string
+  usersFailedToImport: string
+  importPartiallyCompleted: string
+  someUsersProcessed: string
+  importFailedMessage: string
+  importErrors: string
+  downloadErrorReport: string
+  andMoreErrors: string
+  recoverableErrors: string
+  suggestedFix: string
+  canRetry: string
+  andMoreRecoverableErrors: string
+  criticalErrors: string
+  requiresAdminAction: string
+  importAnotherFile: string
+  viewUsers: string
+  nextSteps: string
+  reviewImportedUsers: string
+  verifyManagerRelationships: string
+  testLoginCredentials: string
+  updateAdditionalSettings: string
+  partialSuccessActionRequired: string
+  reviewAndFixErrors: string
+  reimportCorrectedFile: string
+  checkImportedUsers: string
+  considerErrorRecovery: string
+  importErrorReport: string
+  generated: string
+  summaryLabel: string
+  totalErrorsReport: string
+  createdUsersReport: string
+  updatedUsersReport: string
+  failedUsersReport: string
+  errorsLabel: string
+  
+  // Retry Functionality
+  retryThis: string
+  retryAll: string
+  retryingErrors: string
+  autoFixAvailable: string
+  manualFixRequired: string
+  applyAutoFixes: string
+  manuallyFix: string
+  retryComplete: string
+  retryFailed: string
+  retriedRows: string
+  stillFailed: string
+  allErrorsFixed: string
+  someErrorsRemain: string
+  retryInProgress: string
+  processingFixes: string
+
+  // User deletion error messages
+  userNotFound: string
+  cannotDeleteUserManagesEmployees: string
+  cannotDeleteUserHasEvaluations: string
+  cannotDeleteDueToConstraints: string
+  userDeletedSuccessfully: string
+  failedToDeleteUser: string
+  userCreatedSuccessfully: string
+  failedToCreateUser: string
+  userUpdatedSuccessfully: string
+  failedToUpdateUser: string
+  userWithEmailExists: string
+  validationFailed: string
+
+  // Evaluation details for deletion errors
+  evaluationsAsEmployee: string
+  evaluationsAsManager: string
+  evaluationAssignments: string
+  partialAssessments: string
+
+  // Archive functionality
+  archivedEmployees: string
+  viewArchivedEmployees: string
+  archiveEmployee: string
+  archiveSelectedEmployees: string
+  unarchiveEmployee: string
+  archiveReason: string
+  archiveEmployeeTitle: string
+  archiveEmployeeDescription: string
+  archiveReasonPlaceholder: string
+  archiveReasonNote: string
+  archiving: string
+  archivingEmployee: string
+  archivingSelectedEmployees: string
+  close: string
+  unarchivingEmployee: string
+  employeeArchivedSuccessfully: string
+  employeesArchivedSuccessfully: string
+  employeeUnarchivedSuccessfully: string
+  failedToArchiveEmployee: string
+  failedToArchiveEmployees: string
+  failedToUnarchiveEmployee: string
+  cannotArchiveManagerWithActiveReports: string
+  cannotArchiveManager: string
+  cannotArchiveSelf: string
+  activeEmployees: string
+  noDataCanBeDeleted: string
+  archivedOn: string
+  archivedBy: string
+  archiveDetails: string
+  archivedManager: string
+  archivedDepartment: string
+  archivedPosition: string
+  noArchivedEmployees: string
+  noArchivedEmployeesDescription: string
+  noArchivedEmployeesFound: string
+  noArchivedEmployeesFoundDescription: string
+  evaluationHistory: string
+  showEvaluationHistory: string
+  hideEvaluationHistory: string
+  noEvaluations: string
+}
